Keep density consistent with T and P in the added layers for radiation; minor bug fix in oname_coms.
git-svn-id: https://svn.code.sf.net/p/olam-model/code/trunk@765 2bc74c4e-f164-4307-89b2-97abcff1156f

+++ modules/oname_coms.f90
@@ -233,7 +233,7 @@ Module oname_coms
 !!    CO2 PARAMETERS
 
       integer :: co2flag       = 0
-      integer :: co2_ppmv_init = 400.0
+      real    :: co2_ppmv_init = 400.0
       
 !!    SOUNDING SPECIFICATION
